+++ Editor/Common/Enums/PrimitiveMeshType.cs
@@ -2,5 +2,11 @@
 {
     public enum PrimitiveMeshType
     {
+        PLANE,
+        CUBE,
+        UV_SPHERE,
+        ICO_SPHERE,
+        CYLINDER,
+        CAPSULE
     }
 }