+++ Editor/Editors/WorldEditor/RenderSurface/RenderSurfaceView.xaml.cs
@@ -15,8 +15,6 @@ namespace Editor.Editors
     {
         private RenderSurfaceHost _host = null;
         private bool _disposedValue;
-        private bool _canResize = true;
-        private bool _moved = false;
 
         public RenderSurfaceView()
         {
@@ -32,39 +30,6 @@ namespace Editor.Editors
             _host = new RenderSurfaceHost(ActualWidth, ActualHeight);
             _host.MessageHook += new HwndSourceHook(HostMsgFilter);
             Content = _host;
-
-            var window = this.FindVisualParent<Window>();
-
-            Debug.Assert(window is not null);
-
-            var helper = new WindowInteropHelper(window);
-
-            if (helper.Handle != null) HwndSource.FromHwnd(helper.Handle)?.AddHook(HwndMessageHook);
-        }
-
-        private nint HwndMessageHook(nint hwnd, int msg, nint wParam, nint lParam, ref bool handled)
-        {
-            switch ((Win32Msg)msg)
-            {
-                case Win32Msg.WM_SIZING:
-                    _canResize = false;
-                    _moved = false;
-                    break;
-
-                case Win32Msg.WM_ENTERSIZEMOVE:
-                    _moved = true;
-                    break;
-
-                case Win32Msg.WM_EXITSIZEMOVE:
-                    _canResize = true;
-                    if (!_moved) _host.Resize();
-                    break;
-
-                default:
-                    break;
-            }
-
-            return IntPtr.Zero;
         }
 
         private IntPtr HostMsgFilter(IntPtr hwnd, int msg, IntPtr wParam, IntPtr lParam, ref bool handled)
@@ -75,7 +40,6 @@ namespace Editor.Editors
                 case Win32Msg.WM_ENTERSIZEMOVE: throw new Exception();
                 case Win32Msg.WM_EXITSIZEMOVE: throw new Exception();
                 case Win32Msg.WM_SIZE:
-                    if (_canResize) _host.Resize();
                     break;
                 default:
                     break;